# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API de gestion des congrès, soumissions et inscriptions",
        "title": "Congrès API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api",
    "paths": {
...
        "/admin/soumissions": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retourne la liste paginée de toutes les soumissions avec filtres optionnels",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin"
                ],
                "summary": "Lister toutes les soumissions (admin)",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Recherche par titre, auteur, thème ou sujet",
                        "name": "search",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filtrer par type (Abstract, Poster, Communication)",
                        "name": "type",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filtrer par statut (En attente, Approuvée, Rejetée)",
                        "name": "statut",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Numéro de page (défaut: 1)",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Nombre d'éléments par page (défaut: 10)",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/congres-app_backend_pkg_utils.PaginatedResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/congres-app_backend_internal_models.Soumission"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/congres-app_backend_pkg_utils.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/congres-app_backend_pkg_utils.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/soumissions": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retourne la liste des soumissions de l'utilisateur connecté, triées par date de création",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "soumissions"
                ],
                "summary": "Lister les soumissions de l'utilisateur",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/congres-app_backend_pkg_utils.SuccessResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/congres-app_backend_internal_models.Soumission"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/congres-app_backend_pkg_utils.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Soumet un document (PDF) avec les métadonnées associées pour un congrès",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "soumissions"
                ],
                "summary": "Créer une nouvelle soumission",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Type de soumission (Abstract, Poster, Communication)",
                        "name": "submission_type",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Thème",
                        "name": "theme",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Sujets",
                        "name": "topics",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Titre du document",
                        "name": "document_title",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Nom de l'auteur",
                        "name": "author_name",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Résumé",
                        "name": "resume",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Mots-clés (JSON array ou séparés par des virgules)",
                        "name": "keywords",
                        "in": "formData"
                    },
                    {
                        "type": "file",
                        "description": "Fichier PDF",
                        "name": "file",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/congres-app_backend_pkg_utils.SuccessResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/congres-app_backend_internal_models.Soumission"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/congres-app_backend_pkg_utils.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/congres-app_backend_pkg_utils.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "congres-app_backend_internal_models.Notification": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "data": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "id": {
                    "type": "string"
                },
                "notifiable_id": {
                    "type": "string"
                },
                "notifiable_type": {
                    "type": "string"
                },
                "read_at": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "congres-app_backend_internal_models.Soumission": {
            "type": "object",
            "properties": {
                "author_name": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "document_title": {
                    "type": "string"
                },
                "file_path": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "keywords": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "raison_rejet": {
                    "type": "string"
                },
                "resume": {
                    "type": "string"
                },
                "statut": {
                    "description": "En attente, Approuvée, Rejetée",
                    "type": "string"
                },
                "submission_type": {
                    "description": "Abstract, Poster, Communication",
                    "type": "string"
                },
                "theme": {
                    "type": "string"
                },
                "topics": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "user": {
                    "$ref": "#/definitions/congres-app_backend_internal_models.User"
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
        "congres-app_backend_internal_models.User": {
            "type": "object",
            "properties": {
                "active": {
                    "type": "boolean"
                },
                "adresse": {
                    "type": "string"
                },
                "biographie": {
                    "type": "string"
                },
                "civilite": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "nom": {
                    "type": "string"
                },
                "notifications": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/congres-app_backend_internal_models.Notification"
                    }
                },
                "organisme": {
                    "type": "string"
                },
                "prenom": {
                    "type": "string"
                },
                "profession": {
                    "type": "string"
                },
                "role": {
                    "type": "string"
                },
                "sexe": {
                    "type": "string"
                },
                "soumissions": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/congres-app_backend_internal_models.Soumission"
                    }
                },
                "telephone": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "congres-app_backend_pkg_utils.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
        "congres-app_backend_pkg_utils.PaginatedResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "limit": {
                    "type": "integer"
                },
                "page": {
                    "type": "integer"
                },
                "success": {
                    "type": "boolean"
                },
                "total": {
                    "type": "integer"
                }
            }
        },
        "congres-app_backend_pkg_utils.SuccessResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "success": {
                    "type": "boolean"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Bearer JWT token",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}